+++ examples/spk_tgw_gwlb_asg_fgt_igw/main.tf
@@ -74,6 +74,7 @@ locals {
                 nat_gateway_id         = [for k, v in module.ngw : v.nat_gateway.id if v.availability_zone == az][0]
               }
             },
+            existing_rt            = lookup(var.existing_rts, "fgt_login_${az}", null),
             rt_association_subnets = [for k, v in local.subnets : v["id"] if startswith(k, "${local.module_prefix}fgt_login_") && v["availability_zone"] == az]
           }
         },
@@ -86,6 +87,7 @@ locals {
                 gateway_id             = module.security-vpc.igw_id
               }
             },
+            existing_rt            = lookup(var.existing_rts, "ngw_igw", null),
             rt_association_subnets = [for k, v in local.create_subnets : local.subnets["${local.module_prefix}${k}"]["id"] if startswith(k, "ngw_")]
           }
         }
@@ -100,6 +102,7 @@ locals {
               gateway_id             = module.security-vpc.igw_id
             }
           },
+          existing_rt            = lookup(var.existing_rts, "fgt_login", null),
           rt_association_subnets = [for k, v in local.create_subnets : local.subnets["${local.module_prefix}${k}"]["id"] if startswith(k, "fgt_login_")]
         }
       },
@@ -112,6 +115,7 @@ locals {
               transit_gateway_id     = module.transit-gw.tgw.id
             }
           },
+          existing_rt            = lookup(var.existing_rts, "gwlbe", null),
           rt_association_subnets = [for k, v in local.subnets : v["id"] if startswith(k, "${local.module_prefix}gwlbe_")]
         }
       },
@@ -125,6 +129,7 @@ locals {
               vpc_endpoint_id        = [for k, v in local.subnets : module.security-vpc-gwlb.gwlb_endps[k] if startswith(k, "${local.module_prefix}gwlbe_") && v["availability_zone"] == az][0]
             }
           },
+          existing_rt            = lookup(var.existing_rts, "tgw_attachment_${az}", null),
           rt_association_subnets = [for k, v in local.subnets : v["id"] if startswith(k, "${local.module_prefix}tgw_attachment_") && v["availability_zone"] == az]
         }
       }
@@ -172,6 +177,7 @@ module "security_route_table" {
   vpc_id                  = module.security-vpc.vpc_id
   rt_name                 = each.key
   routes                  = lookup(each.value, "routes", {})
+  existing_rt             = lookup(each.value, "existing_rt", null)
   rt_association_subnets  = lookup(each.value, "rt_association_subnets", [])
   rt_association_gateways = lookup(each.value, "rt_association_gateways", [])
   tags = {
@@ -276,6 +282,7 @@ module "fgt_asg" {
   source   = "../../modules/fortigate/fgt_asg"
   for_each = var.asgs
   # FortiGate instance template
+  vpc_id                         = module.security-vpc.vpc_id
   template_name                  = lookup(each.value, "template_name", "")
   ami_id                         = lookup(each.value, "ami_id", null) != null ? each.value.ami_id : lookup(var.fgt_config_shared, "ami_id", null) != null ? var.fgt_config_shared.ami_id : ""
   fgt_version                    = lookup(each.value, "fgt_version", null) != null ? each.value.fgt_version : lookup(var.fgt_config_shared, "fgt_version", null) != null ? var.fgt_config_shared.fgt_version : ""
@@ -312,20 +319,23 @@ module "fgt_asg" {
   metadata_options = lookup(each.value, "metadata_options", null) != null ? each.value.metadata_options : lookup(var.fgt_config_shared, "metadata_options", null)
 
   # Auto Scale Group
-  availability_zones         = var.availability_zones
-  asg_name                   = each.key
-  asg_max_size               = each.value.asg_max_size
-  asg_min_size               = each.value.asg_min_size
-  asg_desired_capacity       = lookup(each.value, "asg_desired_capacity", null)
-  scale_policies             = lookup(each.value, "scale_policies", {})
-  create_dynamodb_table      = lookup(each.value, "create_dynamodb_table", null)
-  dynamodb_table_name        = lookup(each.value, "dynamodb_table_name", null)
-  primary_scalein_protection = lookup(each.value, "primary_scalein_protection", false)
+  availability_zones            = var.availability_zones
+  asg_name                      = each.key
+  asg_max_size                  = each.value.asg_max_size
+  asg_min_size                  = each.value.asg_min_size
+  asg_desired_capacity          = lookup(each.value, "asg_desired_capacity", null)
+  scale_policies                = lookup(each.value, "scale_policies", {})
+  create_dynamodb_table         = lookup(each.value, "create_dynamodb_table", null)
+  dynamodb_table_name           = lookup(each.value, "dynamodb_table_name", null)
+  primary_scalein_protection    = lookup(each.value, "primary_scalein_protection", false)
+  asg_health_check_grace_period = lookup(each.value, "asg_health_check_grace_period", 300)
+  health_check_port             = lookup(var.gwlb_health_check, "port", 0)
+  health_check_protocol         = lookup(var.gwlb_health_check, "protocol", "")
   dynamodb_privatelink = var.enable_privatelink_dydb ? {
     vpc_id                      = module.security-vpc.vpc_id
     region                      = var.region
     privatelink_subnet_ids      = [for k, v in local.subnets : v["id"] if startswith(k, "${local.module_prefix}privatelink_")]
-    privatelink_security_groups = [for sg_name in each.value.privatelink_security_groups : local.secgrp_idmap_with_prefixname["${local.module_prefix}${sg_name}"]]
+    privatelink_security_groups = lookup(each.value, "privatelink_security_groups", null) == null ? [] : [for sg_name in each.value.privatelink_security_groups : local.secgrp_idmap_with_prefixname["${local.module_prefix}${sg_name}"]]
   } : null
   network_interfaces = merge(
     jsondecode(
@@ -492,13 +502,10 @@ module "security-vpc-gwlb" {
   tgp_name                         = var.tgp_name
   deregistration_delay             = 30
   enable_cross_zone_load_balancing = var.enable_cross_zone_load_balancing
-  health_check = {
-    port     = 80
-    protocol = "TCP"
-  }
-  vpc_id               = module.security-vpc.vpc_id
-  gwlb_ln_name         = "gwlb-ln"
-  gwlb_ep_service_name = var.gwlb_ep_service_name
+  health_check                     = var.gwlb_health_check
+  vpc_id                           = module.security-vpc.vpc_id
+  gwlb_ln_name                     = "gwlb-ln"
+  gwlb_ep_service_name             = var.gwlb_ep_service_name
   gwlb_endps = merge([
     { for k, v in local.subnets : k => {
       vpc_id    = module.security-vpc.vpc_id